src/main/java/com/optum/casestudy/service/OptumUserService.java
package com.optum.casestudy.service;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.optum.casestudy.model.User;
import com.optum.casestudy.repository.UserRepository;

@Service
public class OptumUserService {
	@Autowired
	UserRepository userRepository;
	
	public User saveUser(User user) {
		return userRepository.save(user);
	}
	
	public List<User> findAll(){
		return userRepository.findAll();
	}

}
